# docs/swagger.yaml
openapi: 3.0.3
info:
  title: LetRents - Unified Property Management API
  description: |
    **Unified Resource-Based REST API** for LetRents Property Management Platform
    
    ## 🚀 **What's New (API v2.0)**
    This API consolidates **68+ duplicate endpoints** into **45 clean, resource-based endpoints** with:
    - ✅ **Unified RBAC** - Role-based access control across all resources
    - ✅ **Multi-tenancy** - Company-level data isolation 
    - ✅ **Consistent structure** - No more role-based endpoint duplicates
    - ✅ **Better performance** - Reduced API complexity and maintenance overhead
    
    ## 🔐 **Authentication & Authorization**
    All endpoints require Bearer token authentication:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## 🎯 **Role-Based Access Control (RBAC)**
    | Role | Access Level | 
    |------|-------------|
    | **Super Admin** | Full access to all resources across all companies |
    | **Agency Admin** | Full access within their company |
    | **Landlord** | Access to own properties and associated resources |
    | **Agent** | Access to assigned properties/units only |
    | **Caretaker** | Access to assigned properties for maintenance |
    | **Tenant** | Read-only access to own unit and invoices |
    
    ## 🏢 **Multi-Tenancy**
    Data is automatically filtered by `company_id` except for Super Admins.
    
    ## 📊 **Migration from v1.0**
    | Old Endpoints (v1.0) | New Unified Endpoints (v2.0) |
...
    
...
  contact:
...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /properties:
    get:
      tags: [Properties]
      summary: List properties
      description: |
        **Unified endpoint** that replaces:
        - `GET /landlord/properties` 
        - `GET /agency-admin/properties`
        
        Returns properties filtered by user role and company:
        - **Super Admin**: All properties
        - **Agency Admin**: Properties in their company  
        - **Landlord**: Own properties only
        - **Agent**: Assigned properties only
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
        - $ref: '#/components/parameters/Search'
        - name: status
          in: query
          schema:
            type: string
            enum: [active, under_construction, renovation, inactive]
        - name: property_type
          in: query
          schema:
            type: string
            enum: [residential, commercial, industrial, mixed_use]
      responses:
        '200':
          description: Properties retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Properties retrieved successfully"
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Property'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Properties]
      summary: Create property
      description: Create a new property (Agency Admin, Landlord, Super Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePropertyRequest'
      responses:
        '201':
          description: Property created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Property created successfully"
                  data:
                    $ref: '#/components/schemas/Property'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /units:
    get:
      tags: [Units]
      summary: List units
      description: |
        **Unified endpoint** that consolidates multiple unit listing endpoints.
        Returns units filtered by user role and company.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
        - name: property_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by property ID
        - name: status
          in: query
          schema:
            type: string
            enum: [available, occupied, reserved, maintenance, inactive]
        - name: unit_type
          in: query
          schema:
            type: string
            enum: [studio, 1_bedroom, 2_bedroom, 3_bedroom, office_space, retail_shop]
      responses:
        '200':
          description: Units retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Unit'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'

    post:
      tags: [Units]
      summary: Create unit
      description: Create a new unit (Agency Admin, Landlord, Super Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUnitRequest'
      responses:
        '201':
          description: Unit created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Unit created successfully"
                  data:
                    $ref: '#/components/schemas/Unit'
...
  /tenants:
    get:
      tags: [Tenants]
      summary: List tenants
      description: |
        **Unified endpoint** that replaces:
        - `GET /landlord/tenants`
        - `GET /agency-admin/tenants`
        
        Returns tenants filtered by user role and access permissions.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
        - $ref: '#/components/parameters/Search'
        - name: property_id
          in: query
          schema:
            type: string
            format: uuid
        - name: unit_id
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, pending]
      responses:
        '200':
          description: Tenants retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Tenant'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'

    post:
      tags: [Tenants]
      summary: Create tenant
      description: Add a new tenant to the system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTenantRequest'
      responses:
        '201':
          description: Tenant created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Tenant created successfully"
                  data:
                    $ref: '#/components/schemas/Tenant'

  # =============================================================================
  # DASHBOARD (Unified - replaces /landlord/dashboard/* + /agency-admin/dashboard/*)
  # =============================================================================
...
  /notifications:
    get:
      tags: [Notifications]
      summary: Get user notifications
      description: |
        **Unified endpoint** that replaces:
        - `GET /landlord/notifications`
        - `GET /agency-admin/notifications`
        
        Returns notifications for the authenticated user.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
        - name: is_read
          in: query
          schema:
            type: boolean
          description: Filter by read status
        - name: type
          in: query
          schema:
            type: string
            enum: [payment, maintenance, lease, system]
      responses:
        '200':
          description: Notifications retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Notification'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token for authentication
  parameters:
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number for pagination
    PerPage:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
      description: Number of items per page
    Search:
      name: search
      in: query
      schema:
        type: string
      description: Search query for filtering results
  schemas:
...
    Property:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        name:
          type: string
          example: "Westlands Apartments"
        description:
          type: string
          example: "Modern apartment complex in Westlands"
        property_type:
          type: string
          enum: [residential, commercial, industrial, mixed_use]
          example: "residential"
        status:
          type: string
          enum: [active, under_construction, renovation, inactive]
          example: "active"
        total_units:
          type: integer
          example: 24
        occupied_units:
          type: integer
          example: 22
        vacant_units:
          type: integer
          example: 2
        street:
          type: string
          example: "123 Westlands Road"
        city:
          type: string
          example: "Nairobi"
        region:
          type: string
          example: "Nairobi County"
        country:
          type: string
          example: "Kenya"
        postal_code:
          type: string
          example: "00100"
        owner_id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    CreatePropertyRequest:
      type: object
      required: [name, property_type, street, city, country]
      properties:
        name:
          type: string
          example: "Westlands Apartments"
        description:
          type: string
          example: "Modern apartment complex in Westlands"
        property_type:
          type: string
          enum: [residential, commercial, industrial, mixed_use]
          example: "residential"
        street:
          type: string
          example: "123 Westlands Road"
        city:
          type: string
          example: "Nairobi"
        region:
          type: string
          example: "Nairobi County"
        country:
          type: string
          example: "Kenya"
        postal_code:
          type: string
          example: "00100"
        amenities:
          type: array
          items:
            type: string
          example: ["parking", "security", "generator"]
...
    Unit:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174001"
        property_id:
          type: string
          format: uuid
        unit_number:
          type: string
          example: "A1"
        unit_type:
          type: string
          enum: [studio, 1_bedroom, 2_bedroom, 3_bedroom, 4_bedroom, office_space, retail_shop]
          example: "2_bedroom"
        status:
          type: string
          enum: [available, occupied, reserved, maintenance, inactive]
          example: "occupied"
        floor_number:
          type: integer
          example: 2
        size_sqm:
          type: number
          format: decimal
          example: 85.5
        bedrooms:
          type: integer
          example: 2
        bathrooms:
          type: integer
          example: 2
        rent_amount:
          type: number
          format: decimal
          example: 45000.00
        deposit_amount:
          type: number
          format: decimal
          example: 90000.00
        furnishing_type:
          type: string
          enum: [furnished, unfurnished, semi_furnished]
          example: "unfurnished"
        condition:
          type: string
          enum: [new, excellent, good, fair, poor, needs_repairs]
          example: "good"
        tenant_id:
          type: string
          format: uuid
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    CreateUnitRequest:
      type: object
      required: [property_id, unit_number, unit_type, rent_amount]
      properties:
        property_id:
          type: string
          format: uuid
        unit_number:
          type: string
          example: "A1"
        unit_type:
          type: string
          enum: [studio, 1_bedroom, 2_bedroom, 3_bedroom, 4_bedroom, office_space, retail_shop]
          example: "2_bedroom"
        floor_number:
          type: integer
          example: 2
        size_sqm:
          type: number
          format: decimal
          example: 85.5
        bedrooms:
          type: integer
          example: 2
        bathrooms:
          type: integer
          example: 2
        rent_amount:
          type: number
          format: decimal
          example: 45000.00
        deposit_amount:
          type: number
          format: decimal
          example: 90000.00
        furnishing_type:
          type: string
          enum: [furnished, unfurnished, semi_furnished]
          example: "unfurnished"
...
    Tenant:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174002"
        user_id:
          type: string
          format: uuid
        email:
          type: string
          format: email
          example: "tenant@example.com"
        first_name:
          type: string
          example: "Jane"
        last_name:
          type: string
          example: "Smith"
        phone_number:
          type: string
          example: "+254712345679"
        id_number:
          type: string
          example: "12345678"
        unit_id:
          type: string
          format: uuid
        lease_start_date:
          type: string
          format: date
        lease_end_date:
          type: string
          format: date
        monthly_rent:
          type: number
          format: decimal
          example: 45000.00
        deposit_amount:
          type: number
          format: decimal
          example: 90000.00
        status:
          type: string
          enum: [active, inactive, pending]
          example: "active"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateTenantRequest:
      type: object
      required: [email, first_name, last_name, phone_number, unit_id, lease_start_date, monthly_rent]
      properties:
        email:
          type: string
          format: email
        first_name:
          type: string
        last_name:
          type: string
        phone_number:
          type: string
        id_number:
          type: string
        unit_id:
          type: string
          format: uuid
        lease_start_date:
          type: string
          format: date
        lease_end_date:
          type: string
          format: date
        monthly_rent:
          type: number
          format: decimal
        deposit_amount:
          type: number
          format: decimal

    # Dashboard Schemas
...
    Notification:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174003"
        type:
          type: string
          enum: [payment, maintenance, lease, system]
          example: "payment"
        title:
          type: string
          example: "Payment Received"
        message:
          type: string
          example: "Rent payment of KSh 45,000 received for Unit A1"
        is_read:
          type: boolean
          example: false
        data:
          type: object
          description: Additional notification data
        created_at:
          type: string
          format: date-time

    # Message Schema
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 10
        total:
          type: integer
          example: 150
        total_pages:
          type: integer
          example: 15
        has_next_page:
          type: boolean
          example: true
        has_prev_page:
          type: boolean
          example: false
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "An error occurred"
        error:
          type: string
          example: "VALIDATION_ERROR"
        details:
          type: object
          description: Additional error details
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Invalid request data"
            error: "VALIDATION_ERROR"
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Authentication required"
            error: "UNAUTHORIZED"
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Insufficient permissions for this resource"
            error: "FORBIDDEN"
...