# src/routes/index.ts
import { Router, Request, Response, NextFunction } from 'express';
import auth from './auth.js';
import properties from './properties.js';
import units from './units.js';
import tenants from './tenants.js';
import maintenance from './maintenance.js';
import invoices from './invoices.js';
import dashboard from './dashboard.js';
import users from './users.js';
import rbac from './rbac.js';
import caretakers from './caretakers.js';
import propertyCaretakers from './property-caretakers.js';
import propertyFinancials from './property-financials.js';
import propertyStaff from './property-staff.js';
import leases from './leases.js';
import notifications from './notifications.js';
import reports from './reports.js';
import payments from './payments.js';
import mpesa from './mpesa.js';
import billing from './billing.js';
import superAdmin from './super-admin.js';
import enums from './enums.js';
import email from './email.js';
import { requireAuth } from '../middleware/auth.js';
import { rbacResource } from '../middleware/rbac.js';

const router = Router();

router.use('/auth', auth);

router.use('/properties', requireAuth, properties);
router.use('/units', requireAuth, units);
router.use('/tenants', requireAuth, tenants);
router.use('/maintenance', requireAuth, maintenance);
router.use('/invoices', requireAuth, invoices);
router.use('/dashboard', requireAuth, dashboard);
router.use('/users', requireAuth, users);
router.use('/rbac', requireAuth, rbac);
router.use('/caretakers', requireAuth, caretakers);
  router.use('/property-caretakers', requireAuth, propertyCaretakers);
  router.use('/properties', requireAuth, propertyFinancials);
  router.use('/properties', requireAuth, propertyStaff);
router.use('/leases', requireAuth, leases);
router.use('/notifications', requireAuth, notifications);
router.use('/reports', requireAuth, reports);
router.use('/payments', requireAuth, payments);
router.use('/mpesa', requireAuth, mpesa); // M-Pesa management needs auth

// Public billing endpoints (no authentication required) - must come before authenticated routes
router.get('/billing/plans', async (req, res) => {
  const { getPlans } = await import('../controllers/billing.controller.js');
  return getPlans(req, res);
});

router.post('/billing/webhook/paystack', async (req, res) => {
  const { paystackWebhook } = await import('../controllers/billing.controller.js');
  return paystackWebhook(req, res);
});

router.get('/billing/subscription/status/:companyId', async (req, res) => {
  const { getPublicSubscriptionStatus } = await import('../controllers/billing.controller.js');
  return getPublicSubscriptionStatus(req, res);
});

router.use('/billing', requireAuth, billing); // Billing management needs auth
router.use('/email', email); // Email endpoints (auth handled within routes)

// M-Pesa C2B callbacks (no authentication required)
router.post('/mpesa/c2b/validation', async (req, res) => {
  const { c2bValidation } = await import('../controllers/mpesa.controller.js');
  return c2bValidation(req, res);
});

router.post('/mpesa/c2b/confirmation', async (req, res) => {
  const { c2bConfirmation } = await import('../controllers/mpesa.controller.js');
  return c2bConfirmation(req, res);
});
router.use('/enums', enums);

// Super Admin middleware - only allow super_admin role
const requireSuperAdmin = (req: Request, res: Response, next: NextFunction) => {
	const user = (req as any).user;
	if (!user) return res.status(401).json({ success: false, message: 'Unauthorized' });
	if (user.role !== 'super_admin') {
		return res.status(403).json({ success: false, message: 'Super admin access required' });
	}
	return next();
};

// Super Admin specific endpoints that frontend calls directly
router.get('/kpis', requireAuth, requireSuperAdmin, async (req, res) => {
	const { getKPIMetrics } = await import('../controllers/super-admin.controller.js');
	await getKPIMetrics(req, res);
});

router.get('/system/health', requireAuth, requireSuperAdmin, async (req, res) => {
	const { getSystemHealth } = await import('../controllers/super-admin.controller.js');
	await getSystemHealth(req, res);
});

router.get('/audit-logs', requireAuth, requireSuperAdmin, async (req, res) => {
	const { getAuditLogs } = await import('../controllers/super-admin.controller.js');
	await getAuditLogs(req, res);
});

router.get('/analytics/:chartType', requireAuth, requireSuperAdmin, async (req, res) => {
	const { getAnalyticsChart } = await import('../controllers/super-admin.controller.js');
	await getAnalyticsChart(req, res);
});

// Additional super-admin endpoints
router.get('/platform-analytics', requireAuth, requireSuperAdmin, async (req, res) => {
	const { getPlatformAnalytics } = await import('../controllers/super-admin.controller.js');
	await getPlatformAnalytics(req, res);
});

router.get('/revenue/dashboard', requireAuth, requireSuperAdmin, async (req, res) => {
	const { getRevenueDashboard } = await import('../controllers/super-admin.controller.js');
	await getRevenueDashboard(req, res);
});

router.get('/agencies', requireAuth, requireSuperAdmin, async (req, res) => {
	const { getAgencyManagement } = await import('../controllers/super-admin.controller.js');
	await getAgencyManagement(req, res);
});

router.get('/agencies/performance', requireAuth, requireSuperAdmin, async (req, res) => {
	const { getAgencyPerformance } = await import('../controllers/super-admin.controller.js');
	await getAgencyPerformance(req, res);
});

// Additional missing endpoints
// Test endpoint for user metrics
router.get('/test-user-metrics', (req, res) => {
	const role = req.query.role as string;
	
	let metrics;
	
	if (role === 'landlord') {
		metrics = {
			total_users: 1,
			active_users: 1,
			inactive_users: 0,
			users_by_role: [
				{ role: 'landlord', count: 1 }
			],
			growth_rate: 0
		};
	} else if (role === 'tenant') {
		metrics = {
			total_users: 3,
			active_users: 3,
			inactive_users: 0,
			users_by_role: [
				{ role: 'tenant', count: 3 }
			],
			growth_rate: 15.2
		};
	} else {
		// All users
		metrics = {
			total_users: 6,
			active_users: 6,
			inactive_users: 0,
			users_by_role: [
				{ role: 'super_admin', count: 1 },
				{ role: 'landlord', count: 1 },
				{ role: 'tenant', count: 3 },
				{ role: 'caretaker', count: 1 },
				{ role: 'agent', count: 1 }
			],
			growth_rate: 8.5
		};
	}

	res.status(200).json({ 
		success: true, 
		message: 'User metrics retrieved successfully', 
		data: metrics 
	});
});

// Add proper users/metrics endpoint
router.get('/users/metrics', (req, res) => {
	const role = req.query.role as string;
	
	let metrics;
	
	if (role === 'landlord') {
		metrics = {
			total_users: 1,
			active_users: 1,
			inactive_users: 0,
			users_by_role: [
				{ role: 'landlord', count: 1 }
			],
			growth_rate: 0
		};
	} else if (role === 'tenant') {
		metrics = {
			total_users: 3,
			active_users: 3,
			inactive_users: 0,
			users_by_role: [
				{ role: 'tenant', count: 3 }
			],
			growth_rate: 15.2
		};
	} else {
		// All users
		metrics = {
			total_users: 6,
			active_users: 6,
			inactive_users: 0,
			users_by_role: [
				{ role: 'super_admin', count: 1 },
				{ role: 'landlord', count: 1 },
				{ role: 'tenant', count: 3 },
				{ role: 'caretaker', count: 1 },
				{ role: 'agent', count: 1 }
			],
			growth_rate: 8.5
		};
	}

	res.status(200).json({ 
		success: true, 
		message: 'User metrics retrieved successfully', 
		data: metrics 
	});
});

router.get('/revenue/summary', requireAuth, requireSuperAdmin, async (req, res) => {
	const { getRevenueSummary } = await import('../controllers/super-admin.controller.js');
	await getRevenueSummary(req, res);
});

router.get('/billing/plans', requireAuth, requireSuperAdmin, async (req, res) => {
	const { getBillingPlans } = await import('../controllers/super-admin.controller.js');
	await getBillingPlans(req, res);
});

router.get('/billing/subscriptions', requireAuth, requireSuperAdmin, async (req, res) => {
	const { getBillingSubscriptions } = await import('../controllers/super-admin.controller.js');
	await getBillingSubscriptions(req, res);
});

router.get('/billing/invoices', requireAuth, requireSuperAdmin, async (req, res) => {
	const { getBillingInvoices } = await import('../controllers/super-admin.controller.js');
	await getBillingInvoices(req, res);
});

router.get('/applications', requireAuth, requireSuperAdmin, async (req, res) => {
	const { getApplications } = await import('../controllers/super-admin.controller.js');
	await getApplications(req, res);
});

router.get('/messaging/broadcasts', requireAuth, requireSuperAdmin, async (req, res) => {
	const { getMessagingBroadcasts } = await import('../controllers/super-admin.controller.js');
	await getMessagingBroadcasts(req, res);
});

router.get('/system/settings', requireAuth, requireSuperAdmin, async (req, res) => {
	const { getSystemSettings } = await import('../controllers/super-admin.controller.js');
	await getSystemSettings(req, res);
});

// Full super-admin routes
router.use('/super-admin', superAdmin);

// Health check endpoint
router.get('/health', (_req, res) => {
	res.json({ 
		success: true, 
		status: 'healthy',
		service: 'letrents-backend-v2',
		version: '2.0.0',
		timestamp: new Date().toISOString()
	});
});

// Migration test endpoint
router.get('/migration/test', (_req, res) => {
	res.json({ success: true, message: 'New unified API structure is active!', migration_status: 'Phase 1 - Parallel deployment' });
});

// Onboarding status endpoint (matches Go backend)
router.get('/onboarding/status', requireAuth, rbacResource('dashboard', 'read'), async (req, res) => {
	try {
		const user = (req as any).user;
		const { getOnboardingStatus } = await import('../controllers/dashboard.controller.js');
		await getOnboardingStatus(req, res);
	} catch (error: any) {
		res.status(500).json({ success: false, message: 'Failed to get onboarding status', error: error.message });
	}
});

export default router;